// api/src/core/models/Manufacturer/Manufacturer.d.ts
import { Product } from "../Product/Product";
export declare class Manufacturer {
    readonly id: string;
    readonly name: string;
    readonly productByManufacturer: Product[];
    constructor(id: string, name: string, productByManufacturer?: Product[]);
}
//# sourceMappingURL=Manufacturer.d.ts.map